requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.30.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
